websockets-with-fastapi-and-javascript/requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
distro==1.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==1.3.5
openai-whisper==20231117
orjson==3.9.10
playsound==1.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pypiwin32==223
python-dotenv==1.0.0
python-multipart==0.0.6
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
SpeechRecognition==3.10.0
starlette==0.27.0
sympy==1.12
tiktoken==0.5.1
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
whisper==1.1.10
